refactor: move annotation passes into annotate.py (#98 Phase C, step 2)
Second half of the split. The six annotation passes and their exclusive helpers
(20 functions, ~1,980 lines) move out of make_drawing.py into a new
`draftwright/annotate.py`. `_auto_annotate` is the single entry point;
make_drawing's `build_drawing` imports it and nothing else from the new module.

annotate.py imports only from `_core`, `layout`, and third-party libs — never
from make_drawing.py — so the dependency graph stays a DAG:
layout → _core → {make_drawing, annotate}, and make_drawing → annotate.

make_drawing.py drops from 5,267 to 2,934 lines. Orphaned imports (the
third-party names and _core re-exports only the moved passes used) are removed.

Test updates (no behaviour change, only the symbols' module home moved):
- imports of moved functions/constants repointed to `annotate`/`_core`
- the "layout engine wired in" source-grep test now reads annotate.py
- the unfittable-row test monkeypatches the strip solvers on `draftwright.annotate`
  (where _solve_strip_via_layout now looks them up)

CLAUDE.md Architecture updated to the five-module layout (#98 item 4).

Pure move — ruff/mypy clean; full non-slow suite green.
